Create Triplet Generator
This part of code generates the triplet we need for our model to learn in our approach of triplet loss i.e;(Anchor,Positive,Negative) this is the targeted triplet to be generated. This will generate 32 batches of triplet in a single iteration for our model to learn the similar pattern between this images.
This is going to be the heart of our recommendation engine.