Create Deep Learning Model
In this Deep Learning Model we have created 3 functions namely:-triplet_loss which returns us the average loss over the 32 triplets that is 1 batch sent to the model for training.
Next is the get_embedding_network function in which we have defined our pre-trained CNN model. In this we have freezed the base layer in order to preserve the pre-trained values of the model from ImageNet. Then we have added a new Dense Layer which helps in converting our 3D vector to 1D vector.
And in the last we have added get_triplet_model function that returns our model to use which will recommend us the images.