Create Feature Extraction and Indexing.
In this part of code we are give our model all the images that we have to extract their features and test our model. 
We have created a function called as index_all_data that will Collectively store the path to all directories i.e; Collects entire data of images from the database and transfers it to the model.